Refactor BlogPress detail view into modular panels
