+++ src/ui/views/browser/components/blogpress/views/createBackButton.js
@@ -1,0 +1,13 @@
+export default function createBackButton({ onClick = () => {}, label = 'Back to blogs' } = {}) {
+  const button = document.createElement('button');
+  button.type = 'button';
+  button.className = 'blogpress-button blogpress-button--link';
+  button.textContent = label;
+  button.addEventListener('click', event => {
+    event.preventDefault();
+    if (typeof onClick === 'function') {
+      onClick(event);
+    }
+  });
+  return button;
+}
